+++ http/src/main/scala/com/twitter/finatra/http/internal/marshalling/RequestInjectableValues.scala
@@ -44,21 +44,30 @@ class RequestInjectableValues(
 
     if (isRequest(forProperty))
       request
-    else if (hasAnnotation(forProperty, requestParamsAnnotation))
+    else if (hasAnnotation(forProperty, requestParamsAnnotation)) {
       if (forProperty.getType.isCollectionLikeType) {
-        val paramsValue = request.params.getAll(fieldName)
-        if (paramsValue == SeqWithSingleEmptyString && hasStringTypeParam(forProperty))
-          convert(forProperty, Seq())
+        val paramsValue = getCollectionLikePropertyValue(forProperty)
+        if (hasAnnotation[QueryParam](forProperty))
+          queryParamConvert(forProperty, paramsValue)
         else
-          convert(forProperty, request.params.getAll(fieldName))
+          convert(forProperty, paramsValue)
+      } else {
+        request.params.get(fieldName) match {
+          case Some(p) if hasAnnotation[QueryParam](forProperty) =>
+            queryParamConvert(forProperty.getType, p)
+          case Some(p) => convert(forProperty.getType, p)
+          case None => null
+        }
       }
-      else
-        (request.params.get(fieldName) map { convert(forProperty, _) }).orNull
-    else if (hasAnnotation[Header](forProperty))
-      (request.headerMap.get(fieldName) map { convert(forProperty, _) }).orNull
-    else
+    } else if (hasAnnotation[Header](forProperty)) {
+      request.headerMap.get(fieldName) match {
+        case Some(p) => convert(forProperty, p)
+        case None => null
+      }
+    } else {
       injector.getInstance(
         valueId.asInstanceOf[Key[_]]).asInstanceOf[Object]
+    }
   }
 
   /* Private */
@@ -80,6 +89,54 @@ class RequestInjectableValues(
       objectMapper.convert(
         propertyValue,
         forType)
+  }
+
+  private def queryParamConvert(forProperty: BeanProperty, propertyValue: Any): AnyRef = {
+    queryParamConvert(
+      forProperty.getType,
+      propertyValue)
+  }
+
+  private def queryParamConvert(forType: JavaType, propertyValue: Any): AnyRef = {
+    val forTypeClass = forType.getRawClass
+    if (forTypeClass == classOf[Option[_]]) {
+      if (propertyValue == "")
+        None
+      else
+        Option(
+          convert(forType.containedType(0), propertyValue))
+    } else {
+      val modifiedPropertyValue = if (forTypeClass == classOf[java.lang.Boolean]) {
+          matchExtendedBooleans(propertyValue.asInstanceOf[String])
+        } else if (isSeqOfBools(forType)) {
+          propertyValue.asInstanceOf[Seq[String]].map(matchExtendedBooleans)
+        } else {
+          propertyValue
+        }
+
+      objectMapper.convert(
+        modifiedPropertyValue,
+        forType)
+    }
+  }
+
+  private def matchExtendedBooleans(value: String): String = value match {
+    case "t" | "1" => "true"
+    case "f" | "0" => "false"
+    case _ => value
+  }
+
+  private def getCollectionLikePropertyValue(forProperty: BeanProperty): AnyRef = {
+    val fieldName = forProperty.getName
+    if (request.params.getAll(fieldName) == SeqWithSingleEmptyString && hasStringTypeParam(forProperty))
+      Seq.empty
+    else
+      request.params.getAll(fieldName)
+  }
+
+  private def isSeqOfBools(forType: JavaType): Boolean = {
+    forType.getRawClass == classOf[Seq[_]] &&
+      forType.containedType(0).getRawClass == classOf[java.lang.Boolean]
   }
 
   private def hasStringTypeParam(forProperty: BeanProperty): Boolean = {